ethernaut/requirements.txt
dotenv==0.9.9
py-solc-x==2.0.4
pytest==8.4.1
python-dotenv==1.1.0
web3==7.13.0
